change: adds HTML table for drawn cards
