+++ src/routes/+layout.svelte
@@ -16,7 +16,6 @@
 
 <style lang="postcss">
 	.layout {
-		border: 2px solid pink;
 		display: grid;
 		grid-template-columns: 1fr;
 		grid-template-rows: auto 1fr auto;
